requirements.txt
agora_token_builder==1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
asttokens==3.0.0
asyncio==3.4.3
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
av==14.4.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
celery==5.5.3
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.20.0
cmake==3.31.4
colorama==0.4.6
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.0
cssselect2==0.8.0
curl_cffi==0.11.3
cycler==0.12.1
cymem==2.0.11
daphne==4.1.2
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.24
django-channels==0.7.0
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.7
dlib==19.24.6
dnspython==2.7.0
email-validator==2.3.0
env==0.1.0
et_xmlfile==2.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.12
fastapi-mail==1.5.0
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
filelock==3.17.0
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.7.0
future==1.0.0
geographiclib==2.1
geopy==2.4.1
google-crc32c==1.7.1
groq==0.18.0
gunicorn==23.0.0
h11==0.14.0
h3==3.7.7
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.7.1
mplfinance==0.12.10b0
msgpack==1.1.0
multidict==6.6.4
multitasking==0.0.11
murmurhash==1.0.12
mysql-connector==2.2.9
mysql-connector-python==9.0.0
narwhals==1.44.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.3
oauthlib==3.2.2
openai==1.101.0
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.18.1
pillow==10.4.0
platformdirs==4.3.6
plotly==6.1.2
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyJWT==2.10.1
pylibsrtp==0.12.0
pymongo==3.11.4
PyMuPDF==1.26.3
pyOpenSSL==25.0.0
pyparsing==3.2.1
pypdf==5.4.0
pypdfium2==4.30.1
python-barcode==0.15.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
qrcode==8.0
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
routing==0.2.0
rpds-py==0.25.1
seaborn==0.13.2
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
svglib==1.5.1
swisseph==0.0.0.dev1
thinc==8.3.4
timezonefinder==6.5.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.2.3
uvicorn==0.34.0
vine==5.1.0
virtualenv==20.29.2
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
whitenoise==6.11.0
wrapt==1.17.2
xhtml2pdf==0.2.17
yarl==1.20.1
yfinance==0.2.63
yt-dlp==2025.4.30
zope.interface==7.2
